// src/types/supabase.ts
export type Database = {
  public: {
    Tables: {
      session_types: {
        Row: {
          color: string
          created_at: string | null
          default_duration: number
          id: string
          is_flexible: boolean | null
          name: string
          priority: Database["public"]["Enums"]["priority_level"] | null
          updated_at: string | null
        }
        Insert: {
          color: string
          created_at?: string | null
          default_duration: number
          id?: string
          is_flexible?: boolean | null
          name: string
          priority?: Database["public"]["Enums"]["priority_level"] | null
          updated_at?: string | null
        }
        Update: {
          color?: string
          created_at?: string | null
          default_duration?: number
          id?: string
          is_flexible?: boolean | null
          name?: string
          priority?: Database["public"]["Enums"]["priority_level"] | null
          updated_at?: string | null
        }
        Relationships: []
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      priority_level: "low" | "medium" | "high" | "urgent"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type PublicSchema = Database[Extract<keyof Database, "public">]
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
    ? PublicSchema["Enums"][PublicEnumNameOrOptions]
    : never